+++ plugin-dev/Source/Sentry/Public/SentrySubsystem.h
@@ -6,6 +6,7 @@
 #include "Subsystems/EngineSubsystem.h"
 
 #include "SentryDataTypes.h"
+#include "SentryVariant.h"
 #include "SentryScope.h"
 
 #include "SentrySubsystem.generated.h"