+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Type of input parameters in certain public API functions was changed from `FString` to `FSentryVariant`:
   - Get/Set functions for contexts and extras in `SentryScope`
   - Get/Set functions for arbitrary data in `SentryBreadcrumb`
+  - Get/Set functions for arbitrary data in `SentryTransaction` and `SentrySpan`
   - `AddBreadcrumbWithParams` and `SentContext` functions in `SentrySubsystem`
   - `CreateSentryBreadcrumb` function in `SentryLibrary`